Rename unified_tensor module back to `tensor`
Now that the legacy implementation has been removed, we can drop the `unified_`
part of the name, and also combine iterators back into a single module.
